Paralleize writing of DEX code bytes
This will calculate the exact layout of the codes section and then split the
bytes into disjunct sections for writing the bytes in parallel.

This introcudes a DexOutputBuffer with a shared backing array. This type
of DexOutputBuffer ensures that the backing array does not change by
disallowing expanding the backing array.

Bug: b/422947619
Change-Id: I25de84c05579b28da93e7e4160f79ed0d485c254