Only consume feature split program resource providers once
Previously the ClassToFeatureSplitMap was created by reading in all program resources from the program resource providers associated with each feature split. When the provider is backed by a zip this can be rather expensive.

All program resources are read in R8 when building the DexApplication, including the program resources coming from program resource providers associated with feature splits.

This CL updates the feature split program resource providers so that when the program resources are read by the application reader, the program resource providers keep track of the types that have been read.

These sets of types are then used to build the ClassToFeatureSplitMap without disk I/O.

Bug: b/422947619
Change-Id: I781d5c16b26adb13c52a9a95421aa3a100de2055