Rewrite ProguardConfigurationSourceStrings to have String backing
The ProguardConfigurationSourceStrings previously used a List<String> backing, but ProguardConfigurationSourceStrings#get returns a String, by joining the strings in ProguardConfigurationSourceStrings#config.

This CL changes the static type of ProguardConfigurationSourceStrings#config to String. This avoids needing to split the lines of a given string in order to create a ProguardConfigurationSourceStrings, only for the split lines to be joined again when calling ProguardConfigurationSourceStrings#get.

Change-Id: I34d6b2e7d7952fe07152a488f176f0f5349992ae