Stop propagating keep info to companion methods
RELNOTES:

When keep rules match interface methods that are subject to desugaring,
R8 previously transferred the disallow optimization and disallow
shrinking bits to the companion methods.

This changes the behavior so that keep rules no longer apply to
companion methods. This is consistent with the fact that keep rules are
not applicable to other compiler synthesized fields/methods/classes.

Background:

By transferring the disallow optimization and disallow shrinking bits to
the companion methods, the following use case was previously supported:

1) Compile a library with default/static/private interface methods to
DEX with min API < 24 and rules that keep the interface methods. 2)
Compile an app with the library on classpath and -applymapping. 3) Merge
the app and the library.

Note that this only works with -applymapping since the disallow
obfuscation bit is not transferred to the companion methods, i.e., the
companion classes generated from step (1) would have obfuscated method
names.

Going forward the above use case is no longer supported for min API <
24. It is instead possible to do the following:

1) Desugar the library with default/static/private interface methods to
class files with min API < 24. 2) Compile the desugared artifact using
R8 and rules that keep the interface methods on the companion classes.
3) Compile the app with the library on classpath. 4) Merge the app and
the desugared artifact.

Another side effect of this is that it is no longer possible to keep the
inner class and enclosing method attributes for anonymous/local classes
inside interface companion methods.

Change-Id: I7cbe39de0653c4165eb7e7b8274b8dd57f544bcf