Add system property to disable use of one shot resources
AGP 8.x pre-memoizes all ProgramResources and returns the memoized
ProgramResources on each call to
ProgramResourceProvider#getProgramResources. This only works if the
ProgramResources are not one-shot, since R8 may read the app more than
once.

This adds an option to disable the use of one-shot program resources in
R8 to remain compatible with AGP 8.x.

Fixes: b/457998466
Change-Id: Id667dc3bd3701907f9a90302f2cba536fb3226e4