Add releaseFence() to VarHandle desugaring
The implementation will forward to Unsafe.storeFence() (which is what the current libcore implementation does). The desugared API will only
work on API level 24 and above, as this is where Unsafe.storeFence()
was introduced and there are no other ways to simulate this type of
fence.

This CL also adds support for static methods on desugaring classes.

Change-Id: I49681a558e71ff5aa15e03ecfbb8709a31832c63
Bug: b/277462570
Bug: b/457569906