Avoid one shot byte references in deprecated getProgramResources()
When using R8 9 with AGP 8, the AGP provides a custom ProgramResourceProvider that implements the deprecated `Collection<ProgramResource> getProgramResources()`, but not `void getProgramResources(Consumer<ProgramResource>)`.

This provider caches the ProgramResources retrieved from calling `Collection<ProgramResource> ArchiveProgramResourceProvider#getProgramResources()`. This change updates this method to return a collection of non-one-shot ProgramResources, so that it continues to be possible to use an old AGP with a new R8.

Change-Id: Iae40418480c37da64e26fbcfc9413281c7774b2a