Fix pattern evaluation against non-original class in adaptclassstrings
The AdaptClassStringsRewriter and IdentifierMinifier is checking if the
current holder matches one of the -adaptclassstrings patterns. This is
errorprone since the class may have been subject to various
optimizations, such as repackaging. For reliable matching, this CL
performs the matching of -adaptclassstrings up-front and stores the
result in the root set.

Bug: b/462326177
Change-Id: Ic05d6f7fa480b1166cee31f57a44846fdece9a88